refactor: consolidate integration tests into a single Java+Groovy source set
Replaces the three separate integrationTestJunit/Spock/Kotest suites with
a single integrationTest suite backed by test/integration/{java,groovy}/.
Removes Java and Groovy from the unit test source set, leaving only Kotlin/Kotest.
