--- test/integration-kotest/kotlin/testsupport/kotest/ProjectConfig.kt
@@ -1,8 +1,0 @@
-package testsupport.kotest
-
-import io.kotest.core.config.AbstractProjectConfig
-import io.kotest.extensions.decoroutinator.DecoroutinatorExtension
-
-class ProjectConfig : AbstractProjectConfig() {
-  override val extensions = listOf(DecoroutinatorExtension())
-}
