+++ test/integration/groovy/com/mkobit/libraryexample/IntegrationSpec.groovy
